requirements.txt
functions-framework==3.*
google-cloud-eventarc
google-api-python-client
google-auth
firebase-admin
google-cloud-aiplatform
PyMuPDF
